rmeza-SHP-01 Refactored the unit test for the BackEndSerivce
