+++ webapp/js/config.js
@@ -1,12 +1,12 @@
 const config = {
-    "host": "http://localhost",
-    "port": "3000",
+    "host": "",
+    "port": "",
     "rest" : {
-        "size" : "/size",
-        "type" : "/type",
-        "time" : "/time",
-        "transport" : "/transport",
-        "city" : "/cities",
+        "size" : "",
+        "type" : "",
+        "time" : "",
+        "transport" : "",
+        "city" : "",
         "cityPath": "",
         "checkPrice" : "",
         "sendShipping": "",
